Particle effects for buffs and healing, new font, victory scene

+++ Assets/Scripts/GameManager.cs
@@ -1,6 +1,8 @@
+using System;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using UnityEngine.UI;
 using UnityEngine.SceneManagement;
 
 public class GameManager : MonoBehaviour
@@ -8,17 +10,45 @@ public class GameManager : MonoBehaviour
     public bool canRoll, mouseEnabled;
     public string tileColour;
     public GameObject player, sceneObjects, battleObjects, diceObject, menuCanvas, chestCanvas, shopCanvas, buffCanvas, statueBouqet1, statueBouqet2;
+    public Text map1Wins, map2Wins, map3Wins;
+    public int m1wins, m2wins, m3wins;
     PlayerMovement playerM;
     void Start()
-    {
-        playerM = player.GetComponent<PlayerMovement>();
-        if(playerM != null)
+    {        
+        if (player != null)
+        {
+            playerM = player.GetComponent<PlayerMovement>();
+        }
+        //PlayerPrefs.SetString("MapWins", "Map1Map1Map3Map2Map2Map2");
+        foreach (var ch in PlayerPrefs.GetString("MapWins"))
+        {
+            if (ch.ToString().Contains("1"))
+            {
+                m1wins += 1;
+            }
+            else if (ch.ToString().Contains("2"))
+            {
+                m2wins += 1;
+            }
+            else if (ch.ToString().Contains("3"))
+            {
+                m3wins += 1;
+            }
+        }
+        if (playerM != null)
         {
             playerM.showRoll = true;
         }     
     }
     void Update()
-    {      
+    {
+        Scene scene = SceneManager.GetActiveScene();
+        if (player == null && scene.name == "StartScene")
+        {
+            map1Wins.text = m1wins.ToString();
+            map2Wins.text = m2wins.ToString();
+            map3Wins.text = m3wins.ToString();
+        }
         if (Input.GetKeyDown("m") && !menuCanvas.activeSelf && this.GetComponent<TransitionScene>().animationFinished)
         {//Open menu panel
             menuCanvas.SetActive(true);
@@ -32,6 +62,7 @@ public class GameManager : MonoBehaviour
         if (Input.GetMouseButtonUp(1) && canRoll)
         {
             canRoll = false;
+            //GameObject.FindWithTag("Upgrades").GetComponent<ChestScript>().movePicked = false;
         }
         if (PlayerPrefs.HasKey("NextScene") && PlayerPrefs.GetInt("NextScene") == 0)
         {
@@ -42,6 +73,19 @@ public class GameManager : MonoBehaviour
                 canRoll = true;
                 playerM.showRoll = true;
                 PlayerPrefs.DeleteKey("NextScene");
+            }
+            else
+            {//Start the transition into the scene
+                this.GetComponent<TransitionScene>().animateInScene();
+            }
+        }
+        else if (PlayerPrefs.HasKey("NextScene") && PlayerPrefs.GetInt("NextScene") == 3)
+        {
+            if (this.GetComponent<TransitionScene>().animationFinished)
+            {//Show the main scene after losing
+                SceneManager.UnloadSceneAsync("BattleScene");                
+                PlayerPrefs.DeleteKey("NextScene");
+                SceneManager.LoadScene("VictoryScene");
             }
             else
             {//Start the transition into the scene
@@ -78,21 +122,13 @@ public class GameManager : MonoBehaviour
             }
 
         }
-        else if (tileColour.Contains("Yellow"))
+        else if (tileColour.Contains("Yellow") && chestCanvas.activeSelf == false)
         {//Player can add a new attack to their moveset
-            //ChestScene();
+         //ChestScene();           
             Invoke("ChestScene", 2f);
         }
         else if (tileColour.Contains("Blue"))
-        {//Player can receive a permament buff
-            if (tileColour.Contains("1"))
-            {
-                statueBouqet1.SetActive(true);
-            }
-            else if (tileColour.Contains("2"))
-            {
-                statueBouqet2.SetActive(true);
-            }
+        {//Player can choose between a permament buff, hp/mp restore or a revival charge
             BuffScene();
         }
         else if (tileColour.Contains("Black"))
@@ -128,13 +164,24 @@ public class GameManager : MonoBehaviour
     {
         chestCanvas.SetActive(true);
         tileColour = "";
-        PlayerPrefs.DeleteKey("CurrentTile");
+        PlayerPrefs.DeleteKey("CurrentTile");      
     }
     public void BuffScene()
     {
-        buffCanvas.SetActive(true);
-        tileColour = "";
-        PlayerPrefs.DeleteKey("CurrentTile");
+        if (tileColour.Contains("2"))
+        {
+            statueBouqet2.SetActive(true);
+        }
+        else
+        {
+            statueBouqet1.SetActive(true);
+        }
+        if (buffCanvas.activeSelf == false)
+        {
+            buffCanvas.SetActive(true);
+            tileColour = "";
+            PlayerPrefs.DeleteKey("CurrentTile");
+        }
     }
     public void deactivateAll()
     {
